claned up and organized logic into directories and in train.py additional logic to properly save models
